Moved TypeTestCase to it's own namespace
