+++ src/Symfony/Component/Form/Test/TypeTestCase.php
@@ -1,0 +1,42 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien.potencier@symfony-project.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\Form\Test;
+
+use Symfony\Component\Form\FormBuilder;
+use Symfony\Component\Form\Tests\FormIntegrationTestCase;
+use Symfony\Component\EventDispatcher\EventDispatcher;
+
+abstract class TypeTestCase extends FormIntegrationTestCase
+{
+    /**
+     * @var FormBuilder
+     */
+    protected $builder;
+
+    /**
+     * @var EventDispatcher
+     */
+    protected $dispatcher;
+
+    protected function setUp()
+    {
+        parent::setUp();
+
+        $this->dispatcher = $this->getMock('Symfony\Component\EventDispatcher\EventDispatcherInterface');
+        $this->builder = new FormBuilder(null, null, $this->dispatcher, $this->factory);
+    }
+
+    public static function assertDateTimeEquals(\DateTime $expected, \DateTime $actual)
+    {
+        self::assertEquals($expected->format('c'), $actual->format('c'));
+    }
+}
