+++ src/Symfony/Component/Form/Tests/Extension/Core/Type/CollectionTypeTest.php
@@ -13,7 +13,7 @@ namespace Symfony\Component\Form\Tests\Extension\Core\Type;
 
 use Symfony\Component\Form\Form;
 
-class CollectionTypeTest extends TypeTestCase
+class CollectionTypeTest extends \Symfony\Component\Form\Test\TypeTestCase
 {
     public function testContainsNoChildByDefault()
     {
