+++ src/Symfony/Component/Form/Tests/Extension/Core/Type/TimezoneTypeTest.php
@@ -13,7 +13,7 @@ namespace Symfony\Component\Form\Tests\Extension\Core\Type;
 
 use Symfony\Component\Form\Extension\Core\View\ChoiceView;
 
-class TimezoneTypeTest extends TypeTestCase
+class TimezoneTypeTest extends \Symfony\Component\Form\Test\TypeTestCase
 {
     public function testTimezonesAreSelectable()
     {
